requirements.txt
Django==1.11.7
django-bootstrap4==0.0.4
django-rest-framework==0.1.0
djangorestframework==3.7.3
pytz==2017.3
